refactor: 投稿回答の検証を型へ移す
回答投稿時の整合性検証を PostedAnswerContents に閉じ込め、未検証の回答集合から AnswerEntry を生成できないようにした。

タイトル生成も検証済み回答と取得済み質問を再利用する流れへ合わせ、複数選択の互換パースと doctest を維持した。
