+++ server/domain/Cargo.toml
@@ -11,6 +11,7 @@ deriving_via = { workspace = true }
 errors = { path = "../errors" }
 mockall = { workspace = true }
 serde = { workspace = true }
+serde_json = { workspace = true }
 strum = { workspace = true }
 strum_macros = { workspace = true }
 types = { path = "../types" }
@@ -24,6 +25,5 @@ common = { path = "../common" }
 chrono = { workspace = true, features = ["arbitrary"] }
 proptest = { workspace = true }
 proptest-derive = { workspace = true }
-serde_json = { workspace = true }
 test-case = { workspace = true }
 types = { path = "../types", features = ["arbitrary"] }
